add linearized iterator
- add linearized iterator
- provide interface to select itertor traversing policy
- move `makeIter ` into `onAcc` namespace
