+++ include/alpaka/onAcc.hpp
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include "alpaka/mem/Iter.hpp"
 #include "alpaka/mem/MdSpan.hpp"
 #include "alpaka/onAcc/internal.hpp"
 
@@ -26,5 +27,66 @@ namespace alpaka::onAcc
         using CArrayType = typename CArrayType<T, T_Extent>::type;
         return MdSpanArray<CArrayType>{internalCompute::declareSharedVar<CArrayType>(acc)};
     }
+
+    /**
+     * ALPAKA_FN_HOST_ACC is required for cuda else __host__ function called from __host__ __device__ warning
+     * is popping up and generated code is wrong.
+     * @{
+     */
+    template<
+        iter::concepts::IdxTraversing T_Traverse = iter::traverse::Linearized,
+        iter::concepts::IdxMapping T_IdxMapping = iter::layout::Optimize>
+    ALPAKA_FN_HOST_ACC constexpr auto makeIter(
+        auto const& acc,
+        auto rangeOps,
+        T_Traverse traverse = T_Traverse{},
+        T_IdxMapping idxMapping = T_IdxMapping{})
+    {
+        return iter::internal::MakeIter::Op<ALPAKA_TYPE(acc), ALPAKA_TYPE(rangeOps), T_Traverse, T_IdxMapping>{}(
+            acc,
+            rangeOps,
+            traverse,
+            idxMapping);
+    }
+
+    template<
+        iter::concepts::IdxTraversing T_Traverse = iter::traverse::Linearized,
+        iter::concepts::IdxMapping T_IdxMapping = iter::layout::Optimize>
+    ALPAKA_FN_HOST_ACC constexpr auto makeIter(
+        auto const& acc,
+        auto rangeOps,
+        alpaka::concepts::Vector auto const& extent,
+        T_Traverse traverse = T_Traverse{},
+        T_IdxMapping idxMapping = T_IdxMapping{})
+    {
+        return iter::internal::MakeIter::Op<ALPAKA_TYPE(acc), ALPAKA_TYPE(rangeOps), T_Traverse, T_IdxMapping>{}(
+            acc,
+            rangeOps,
+            traverse,
+            idxMapping,
+            extent);
+    }
+
+    template<
+        iter::concepts::IdxTraversing T_Traverse = iter::traverse::Linearized,
+        iter::concepts::IdxMapping T_IdxMapping = iter::layout::Optimize>
+    ALPAKA_FN_HOST_ACC constexpr auto makeIter(
+        auto const& acc,
+        auto rangeOps,
+        alpaka::concepts::Vector auto const& offset,
+        alpaka::concepts::Vector auto const& extent,
+        T_Traverse traverse = T_Traverse{},
+        T_IdxMapping idxMapping = T_IdxMapping{})
+    {
+        return iter::internal::MakeIter::Op<ALPAKA_TYPE(acc), ALPAKA_TYPE(rangeOps), T_Traverse, T_IdxMapping>{}(
+            acc,
+            rangeOps,
+            traverse,
+            idxMapping,
+            offset,
+            extent);
+    }
+
+    /** @} */
 
 } // namespace alpaka::onAcc
